+++ app/mailers/dorsale/flyboy/task_mailer.rb
@@ -9,6 +9,14 @@ module Dorsale
           :subject => t("emails.task.new.subject"),
         )
       end
+
+      def term_email(task)
+        @task = task
+        mail(
+          :to      => task.owner.email,
+          :subject => t("emails.task.term.subject", task = task.name),
+        )
+      end
     end
   end
 end