+++ config/schedule.rb
@@ -1,0 +1,5 @@
+if ENV["RAILS_ENV"] == "production"
+  every :day, at: "7 am" do
+    rake "cron:daily"
+  end
+end